% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{discus}
\alias{discus}
\title{Storm Discussions}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12847 rows and 6 columns.}
\usage{
discus
}
\description{
\describe{
  \item{Status}{Classification of storm, e.g., Tropical Storm, Hurricane,
    etc.}
  \item{Name}{Name of storm}
  \item{Adv}{Advisory Number}
  \item{Date}{Date of advisory issuance}
  \item{Key}{ID of cyclone}
  \item{Contents}{Text content of product}
}
}
\keyword{datasets}
