% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fcst}
\alias{fcst}
\title{Forecasts - Forecast tidy data from \code{fstadv}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 66170 rows and 8 columns.}
\usage{
fcst
}
\description{
\describe{
   \item{Key}{Unique identifier of cyclone}
   \item{Adv}{Advisory number}
   \item{Date}{Date and time of advisory}
   \item{FcstDate}{Forecast date and time in UTC}
   \item{Lat}{Forecast latitude}
   \item{Lon}{Forecast Longitude}
   \item{Wind}{Forecast wind in knots}
   \item{Gust}{Forecast gust in knots}
}
}
\keyword{datasets}
