% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fcst_wr}
\alias{fcst_wr}
\title{Forecast Wind Radius - Forecast wind radius tidied from \code{fstadv}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 85255 rows and 9 columns.}
\usage{
fcst_wr
}
\description{
\describe{
   \item{Key}{Unique identifier of cyclone}
   \item{Adv}{Advisory number}
   \item{Date}{Date and time of advisory}
   \item{FcstDate}{Forecast date and time in UTC}
   \item{WindField}{Minimum sustained wind field for quadrants}
   \item{NE}{Radius in nautical miles for northeast quadrant}
   \item{SE}{Radius in nautical miles for southeast quadrant}
   \item{SW}{Radius in nautical miles for southwest quadrant}
   \item{NW}{Radius in nautical miles for northwest quadrant}
}
}
\keyword{datasets}
