% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fstadv}
\alias{fstadv}
\title{Forecast/Advisory}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12874 rows and 117 columns.}
\usage{
fstadv
}
\description{
\describe{
   \item{Status}{Classification of cyclone}
   \item{Name}{Name of cyclone}
   \item{Adv}{Advisory number}
   \item{Date}{Date and time of advisory}
   \item{Key}{Unique identifier of cyclone}
   \item{Lat}{Latitude of cyclone center}
   \item{Lon}{Longitude of cyclone center}
   \item{Wind}{Maximum sustained one-minute winds in knots}
   \item{Gust}{Maximum sustained one-minute gusts in knots}
   \item{Pressure}{Minimum central pressure in millibars}
   \item{PosAcc}{Position accuracy of cyclone in nautical miles}
   \item{FwdDir}{Compass angle of forward motion}
   \item{FwdSpeed}{Forward speed in miles per hour}
   \item{Eye}{Size of eye in nautical miles}
   \item{NE64}{Radius of >=64kt winds in northeast quadrant}
   \item{SE64}{Radius of >=64kt winds in southeast quadrant}
   \item{SW64}{Radius of >=64kt winds in southwest quadrant}
   \item{NW64}{Radius of >=64kt winds in northwest quadrant}
   \item{NE50}{Radius of >=50kt winds in northeast quadrant}
   \item{SE50}{Radius of >=50kt winds in southeast quadrant}
   \item{SW50}{Radius of >=50kt winds in southwest quadrant}
   \item{NW50}{Radius of >=50kt winds in northwest quadrant}
   \item{NE34}{Radius of >=34kt winds in northwest quadrant}
   \item{SE34}{Radius of >=34kt winds in southeast quadrant}
   \item{SW34}{Radius of >=34kt winds in southwest quadrant}
   \item{NW34}{Radius of >=34kt winds in northwest quadrant}
   \item{Hr\{n\}FcstDate}{Forecast valid date}
   \item{Hr\{n\}Lat}{Forecast latitude in `n` hours}
   \item{Hr\{n\}Lon}{Forecast longitude in `n` hours}
   \item{Hr\{n\}Wind}{Forecast maximum wind in `n` hours}
   \item{Hr\{n\}Gust}{Forecast maximum gust in `n` hours}
   \item{Hr\{n\}NE64}{Forecast wind radius in `n` hours}
   \item{Hr\{n\}SE64}{Forecast wind radius in `n` hours}
   \item{Hr\{n\}SW64}{Forecast wind radius in `n` hours}
   \item{Hr\{n\}NW64}{Forecast wind radius in `n` hours}
   \item{Hr\{n\}NE50}{Forecast wind radius in `n` hours}
   \item{Hr\{n\}SE50}{Forecast wind radius in `n` hours}
   \item{Hr\{n\}SW50}{Forecast wind radius in `n` hours}
   \item{Hr\{n\}NW50}{Forecast wind radius in `n` hours}
   \item{Hr\{n\}NE34}{Forecast wind radius in `n` hours}
   \item{Hr\{n\}SE34}{Forecast wind radius in `n` hours}
   \item{Hr\{n\}SW34}{Forecast wind radius in `n` hours}
   \item{Hr\{n\}NW34}{Forecast wind radius in `n` hours}
   \item{SeasNE}{Radius of 12ft seas in northeast quadrant}
   \item{SeasSE}{Radius of 12ft seas in southeast quadrant}
   \item{SeasSW}{Radius of 12ft seas in southwest quadrant}
   \item{SeasNW}{Radius of 12ft seas in northwest quadrant}
}
}
\keyword{datasets}
