% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{posest}
\alias{posest}
\title{Position Estimates}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 85 rows and 4 columns.}
\usage{
posest
}
\description{
\describe{
  \item{Status}{Classification of storm, e.g., Tropical Storm, Hurricane,
    etc.}
  \item{Name}{Name of storm}
  \item{Date}{Date of advisory issuance}
  \item{Contents}{Text content of product}
}
}
\keyword{datasets}
