% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prblty}
\alias{prblty}
\title{Strike Probabilities}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 55571 rows and 10 columns.}
\usage{
prblty
}
\description{
\describe{
  \item{Status}{Classification of storm, e.g., Tropical Storm, Hurricane,
    etc.}
  \item{Name}{Name of storm}
  \item{Adv}{Advisory Number}
  \item{Date}{Date of advisory issuance}
  \item{Location}{Location for which the probability statistics rely}
  \item{A}{Probability of a strike within the next 12 hours}
  \item{B}{Probability of a strike between 12 and 24 hours}
  \item{C}{Probability of a strike between 24 and 36 hours}
  \item{D}{Probability of a strike between 36 and 48 hours}
  \item{E}{Probability of a strike between 48 and 72 hours}
}
}
\keyword{datasets}
