% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{public}
\alias{public}
\title{Public Advisory}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14080 rows and 6 columns.}
\usage{
public
}
\description{
\describe{
  \item{Status}{Classification of storm, e.g., Tropical Storm, Hurricane,
    etc.}
  \item{Name}{Name of storm}
  \item{Adv}{Advisory Number}
  \item{Date}{Date of advisory issuance}
  \item{Key}{Unique ID of the cyclone}
  \item{Contents}{Text content of product}
}
}
\keyword{datasets}
