% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{rrricanesdata}
\alias{rrricanesdata}
\alias{rrricanesdata-package}
\title{rrricanesdata}
\description{
rrricanesdata is the complimentary data package to `rrricanes`.
  It includes all product datasets for north Atlantic and northeast
  Pacific tropical cyclones from 1998 through the previous month. This
  package will be updated on the first of every month provided at least one
  cyclone advisory was issued in the previous month.
}
