% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{storms}
\alias{storms}
\title{Storm Summary}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 647 rows and 5 columns.}
\usage{
storms
}
\description{
\describe{
  \item{Key}{Storm ID}
  \item{Name}{Storm name}
  \item{Wind}{Peak wind speed in knots}
  \item{StartDate}{Date/time of first advisory}
  \item{EndDate}{Date/time of last advisory}
}
}
\keyword{datasets}
