% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{update}
\alias{update}
\title{Tropical Cyclone Updates}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 237 rows and 5 columns.}
\usage{
update
}
\description{
\describe{
  \item{Status}{Classification of storm, e.g., Tropical Storm, Hurricane,
    etc.}
  \item{Name}{Name of storm}
  \item{Date}{Date of advisory issuance}
  \item{Key}{Unique ID of cyclone}
  \item{Contents}{Text content of product}
}
}
\keyword{datasets}
