% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wndprb}
\alias{wndprb}
\title{Wind Speed Probabilities}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 118222 rows and 18 columns.}
\usage{
wndprb
}
\description{
\describe{
  \item{Status}{Classification of storm, e.g., Tropical Storm, Hurricane,
    etc.}
  \item{Name}{Name of storm}
  \item{Adv}{Advisory Number}
  \item{Date}{Date of advisory issuance}
  \item{Wind}{Minimum wind speed for which probabilities reference}
  \item{Wind12}{Probability of sustained `Wind` within 12 hours}
  \item{Wind24}{Probability of sustained `Wind` within 24 hours}
  \item{Wind24Cum}{Cumulative probability through 24 hours}
  \item{Wind36}{Probability of sustained `Wind` within 36 hours}
  \item{Wind36Cum}{Cumulative probability through 36 hours}
  \item{Wind48}{Probability of sustained `Wind` within 48 hours}
  \item{Wind48Cum}{Cumulative probability through 48 hours}
  \item{Wind72}{Probability of sustained `Wind` within 72 hours}
  \item{Wind72Cum}{Cumulative probability through 72 hours}
  \item{Wind96}{Probability of sustained `Wind` within 96 hours}
  \item{Wind96Cum}{Cumulative probability through 96 hours}
  \item{Wind120}{Probability of sustained `Wind` within 120 hours}
  \item{Wind120Cum}{Cumulative probability through 120 hours}
}
}
\keyword{datasets}
