% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wr}
\alias{wr}
\title{Wind Radius}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19294 rows and 8 columns.}
\usage{
wr
}
\description{
\describe{
   \item{Key}{Unique identifier of cyclone}
   \item{Adv}{Advisory number}
   \item{Date}{Date and time of advisory}
   \item{Windfield}{Minimum wind speed expected}
   \item{NE}{Radius of `Windfield` in the northeast quadrant}
   \item{SE}{Radius of `Windfield` in the southeast quadrant}
   \item{SW}{Radius of `Windfield` in the southwest quadrant}
   \item{NW}{Radius of `Windfield` in the northwest quadrant}
}
}
\keyword{datasets}
